Renamed noise-pass to noise
